requirements.txt
fastapi
uvicorn[standard]
httpx[http2]
beautifulsoup4
lxml
sentence-transformers
scikit-learn
python-dotenv
pydantic
asyncio
httpx
requests 